ai-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
Pillow==10.3.0
numpy==1.24.4
python-dotenv==1.0.1

# TensorFlow — uncomment one line based on your Python version:
# Python 3.8:
# tensorflow==2.13.0
# Python 3.10+:
# tensorflow==2.16.1
# Or install CPU-only for faster installs without GPU:
# tensorflow-cpu==2.16.1
